MAGE-1534: introduce AbstractClientProvider class
